pkg/containerfs: move to internal
The only external consumer are the `graphdriver` and `graphdriver/shim`
packages in github.com/docker/go-plugins-helpers, which depended on
[ContainerFS][1], which was removed in 9ce2b30b817ad6fe6da8d91e5ac0aa19fb64e0d1.

graphdriver-plugins were deprecated in 6da604aa6a74ab770207932b3cbc3d009c3ae25f,
and support for them removed in 555dac5e14ad4925e020f1a72e8c39b7a316b0dc,
so removing this should not be an issue.

Ideally this package would've been moved inside `daemon/internal`, but it's used
by the `daemon` (cleanupContainer), `plugin` package, and by `graphdrivers`,
so needs to be in the top-level `internal/` package.

[1]: https://github.com/docker/go-plugins-helpers/blob/6eecb7beb65124bb44a23848bb46e98b4f50ae18/graphdriver/api.go#L218
