+++ internal/containerfs/rm.go
