+++ internal/containerfs/rm_test.go
