+++ internal/containerfs/rm_windows.go
